import_usb.py
import usb.core
import usb.util

# Define the Vendor ID and Product ID of your USB device (replace these with the correct values)
VENDOR_ID = 0x1234  # Example Vendor ID
PRODUCT_ID = 0x5678  # Example Product ID

# This is the secret key stored on the USB device. In practice, this could be more secure.
SECRET_KEY = "my_secret_key"  # Replace with your own secret key or token

# Find the USB device
dev = usb.core.find(idVendor=VENDOR_ID, idProduct=PRODUCT_ID)

if dev is None:
    raise ValueError('Device not found')

# Check if the device is connected
dev.set_configuration()

# Read data from the USB device (assuming it's stored in a readable format)
# This step will depend on your device's configuration, but for a simple example, let's read a string
# from an endpoint (usually the device will expose one or more endpoints to communicate with).

# You may need to adjust this based on how your USB device is configured.
endpoint = dev[0][(0,0)][0]  # Using the first endpoint of the device

try:
    data = dev.read(endpoint.bEndpointAddress, endpoint.wMaxPacketSize)
    print(f"Data from USB device: {data}")
    
    # Here, you would compare the data from the device with the SECRET_KEY
    if SECRET_KEY in data.decode():
        print("Authentication successful: USB device verified!")
    else:
        print("Authentication failed: Key mismatch.")
except usb.core.USBError as e:
    print(f"Error reading from USB device: {e}")
